refactor(migration): 마이그레이션 공유 상수/표시 라벨 단일 소스로 통합
- IssueType 표시 라벨 5중복 dict를 migration_constants.ISSUE_TYPE_DISPLAY_NAMES로 통합
- AUTO_FIXABLE_TYPES 2중복을 migration_constants.AUTO_FIXABLE_ISSUE_TYPES로 통합
- MySQL 8.4 예약어 목록 2중복(MigrationAnalyzer.NEW_RESERVED_KEYWORDS ↔ ALL_RESERVED_KEYWORDS) 제거
- 식별자 컨텍스트 매처(_IdentifierIssuePattern 등)를 migration_identifier_matchers.py로 분리하고 migration_constants.py에서 re-export
- STORAGE_ENGINE_STATUS 타입 어노테이션 오타(Dict[str, any] → Dict[str, Any]) 수정

순수 behavior-preserving 리팩터. 로직/제어흐름 변경 없음.
